requirements.txt
beautifulsoup4==4.12.3
cachetools==5.5.2
certifi==2026.2.25
charset-normalizer==3.4.6
cloudscraper==1.2.71
google-auth==2.30.0
google-auth-oauthlib==1.3.0
gspread==6.1.2
idna==3.11
lxml==6.0.2
oauthlib==3.3.1
psycopg==3.2.10
psycopg-binary==3.2.10
pyasn1==0.6.3
pyasn1_modules==0.4.2
pyparsing==3.3.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9.1
soupsieve==2.8.3
tzdata==2025.3
urllib3==2.6.3
